♻️ api함수 목적에 맞추어 분리
