+++ src/api/questions.js
@@ -1,0 +1,13 @@
+import axios from "./axios";
+
+export const getQuestionList = async ({ subjectId, limit = 8, offset = 0 }) => {
+  try {
+    const response = await axios.get(
+      `subjects/${subjectId}/questions/?limit=${limit}&offset=${offset}`,
+    );
+    return response.data;
+  } catch (error) {
+    console.error(error.message);
+    throw error;
+  }
+};
